+++ changelog.rst
@@ -8,3 +8,8 @@ added type hints, type checking with mypy fix importing issues,
 clean __init__ files
 pyaqsapi now builds both pypi and conda packages
 pyaqsapi is now available via pypi
+
+.v1.0.2::
+fix issue that caused aqs_remove_header function to throw an error
+replace sphinxcontrib/apidoc with sphinx.ext.apidoc
+remove *test directories from documentation